Change 64-bit ID to 32-bit; Fix code for WeldRecipe

+++ L20_RTP/AlarmData.h
@@ -24,7 +24,7 @@ struct ALARM_DATA
 {
 	unsigned int 	AlarmType;
 	unsigned int	AlarmSubType;
-	long long		WeldResultID;
+	long			WeldResultID;
 	unsigned int	WeldRecipeID;
 	unsigned int	UserID;
 };
